This is the final commit for scrape_bing.py as well as the removal of BingCovid.py
